+++ CHANGES.rst
@@ -288,6 +288,8 @@ astropy.wcs
 Other Changes and Additions
 ---------------------------
 
+- Minimum version of supported Numpy is now 1.17. [#10664]
+
 - The private ``_erfa`` module has been converted to its own package,
   ``pyerfa``, which is a required dependency for astropy, and can be imported
   with ``import erfa``.  Importing ``_erfa`` from ``astropy`` will give a